Bug corrected, functions added
dsk-lua had problems when saving a big file, this is now corrected
Added the function "savewithcustomheader" which lets you write a custom header for your files. A nice way to add some code into it when room on the disk is scarce. A quick explanation is given in an article I wrote some time ago at https://thecheshirec.at/2023/10/23/mettre-du-code-dans-un-header-amdos/ (unfortunately in French, sorry)