Errors while opening a DSK created by RVM2
RVM2 says there is 43 tracks in its DSK but the fact is there is only 40.
I tried to make a generic correction : dsklua now counts in the "track size table" the number of tracks listed. If the number is different, the number of tracks is replaced by the effective number found in the track size table.